Use hostname and IP address to identify clients
This fix allows clients with duplicate hostnames to work correctly.
If there was two clients with the same hostname then the server was
assuming they were running on the same computer and would only send
setup and project files to a single client.

This fix will allow Amleto to run on AWS correctly.